# [WIP][PoC][testing] Python array.array interface