--- a/python/setup.py.in
+++ b/python/setup.py.in
@@ -117,6 +117,11 @@ extra_link_args = list(ldconf)
 if platform.system() == 'Darwin':
     extra_link_args.append('-Wl,-rpath,'+"${CMAKE_INSTALL_PREFIX}/${CMAKE_INSTALL_LIBDIR}")
 
+if platform.system() == 'Windows':
+    libname = "cryptominisat5win"
+else:
+    libname = "cryptominisat5"
+
 modules = dict(
     name = "pycryptosat",
     sources = ["${CMAKE_CURRENT_SOURCE_DIR}/src/pycryptosat.cpp"],
@@ -126,7 +131,7 @@ modules = dict(
     language = "c++",
     library_dirs=['.', '${PROJECT_BINARY_DIR}/lib'],
     runtime_library_dirs=['${CMAKE_INSTALL_PREFIX}/${CMAKE_INSTALL_LIBDIR}'],
-    libraries = ['cryptominisat5']
+    libraries = ['libname']
 )
 
 setup(