--- a/python/src/pycryptosat.cpp
+++ b/python/src/pycryptosat.cpp
@@ -471,7 +471,9 @@ static PyObject* add_clauses(Solver *self, PyObject *args, PyObject *kwds)
     while ((clause = PyIter_Next(iterator)) != NULL) {
         PyObject *arglist = Py_BuildValue("(O)", clause);
         PyObject *ret = add_clause(self, arglist, NULL);
-        Py_DECREF(ret);
+        if (ret) {
+            Py_DECREF(ret);
+        }
 
         /* release reference when done */
         Py_DECREF(arglist);