… src/comphandler.h
… src/cryptominisat.cpp
… src/cryptominisat.h.in
… src/occsimplifier.cpp
… src/solver.cpp
… src/solver.h