\[ DISCLAIMER: This is no production-ready code, just for testing and unfinished, plus I know next to nothing about how to efficiently work with Python C-API :wink: \]

This is in reference to
- https://github.com/msoos/cryptominisat/issues/513, and
- https://github.com/conda/conda/pull/7676

I put together some non-polished interface to be able to pass clauses to `Solver.add_clauses` as a Python `array.array("i")` object. This is so we can compare the time taken on processing the clauses as a simple stream of integers (separated by zeros) to processing them as iterables of iterables, e.g., list of tuples, which the current interface expects.
Since the time taken to scan an integer array is practically zero (if we need to see all values anyways, that is), the remaining time of adding clauses can be assumed to be spent outside of the Python interface, i.e., be admeasured to the core library.

I'll follow up with some measurement results..