@jakirkham 2018-08-24
So I haven't dug into this code at all, which means my comments could be totally irrelevant and/or addressed. Just came across it while reading issue ( https://github.com/msoos/cryptominisat/issues/513 ).

Not sure if you are aware (or how much it matters for this use case), but Python has a [buffer protocol]( https://docs.python.org/3/c-api/buffer.html ) (supported in both Python 2 and Python 3), which allows zero copy sharing of data between different Python representations  (e.g. `array.array`, `bytes`, `numpy.ndarray`, etc.). This can also be used to wrap `std::vector` or other C/C++ array objects and have them representable in Python cheaply. The buffer protocol can be implemented in [C/C++]( https://jakevdp.github.io/blog/2014/05/05/introduction-to-the-python-buffer-protocol/ ). Also can leverage utilities in [Cython]( https://cython.readthedocs.io/en/latest/src/userguide/buffer.html ), [pybind11]( https://pybind11.readthedocs.io/en/stable/advanced/pycpp/numpy.html#buffer-protocol ), [Boost]( https://stackoverflow.com/a/17759608 ), etc. Happy to answer questions if it helps.