@mbargull 2018-08-24
The following is in reference to  and shows `conda`-specific results:

----
*Before* merging with the recent improvements on [`master`](https://github.com/msoos/cryptominisat/issues/513#issuecomment-415556691) and with `5.6.4` as a base we get the following results:

```sh
$ export CONDA_CHANNELS=defaults CONDA_INSTRUMENTATION_ENABLED=1 CONDA_TIMED_LOGGING=1

$ # 5.6.4 + python-array + ClauseArray
$ conda create -nx --dry-run -v anaconda
Solving environment: ...working... === time_recorder total time and calls ===
TOTAL   116.266       853 conda.common.logic.Clauses.LinearBound_
TOTAL   104.667        62 conda.common.logic.Clauses._run_sat
TOTAL   217.304        12 conda.common.logic.Clauses.minimize
TOTAL   104.670        62 conda.common.logic.Clauses.sat
TOTAL     7.272        62 conda.common.logic.CryptoMiniSatSolver.invoke
TOTAL     1.641        62 conda.common.logic.CryptoMiniSatSolver.process_solution
TOTAL    62.973        62 conda.common.logic.CryptoMiniSatSolver.setup
TOTAL    24.026        62 conda.common.logic.PycoSatSolver.setup
TOTAL    75.510        62 conda.common.logic.SatSolver.run
TOTAL    20.790         2 conda.resolve.Resolve.gen_clauses
TOTAL    13.895         2 conda.resolve.Resolve.get_reduced_index
TOTAL   232.203         1 conda.resolve.Resolve.solve

$ # 5.6.4 + python-array + ClauseList
$ conda create -nx --dry-run -v anaconda
Solving environment: ...working... === time_recorder total time and calls ===
TOTAL    79.806       853 conda.common.logic.Clauses.LinearBound_
TOTAL   176.648        62 conda.common.logic.Clauses._run_sat
TOTAL   253.282        12 conda.common.logic.Clauses.minimize
TOTAL   176.652        62 conda.common.logic.Clauses.sat
TOTAL     7.336        62 conda.common.logic.CryptoMiniSatSolver.invoke
TOTAL     1.652        62 conda.common.logic.CryptoMiniSatSolver.process_solution
TOTAL   119.002        62 conda.common.logic.CryptoMiniSatSolver.setup
TOTAL    39.833        62 conda.common.logic.PycoSatSolver.setup
TOTAL   131.348        62 conda.common.logic.SatSolver.run
TOTAL    20.426         2 conda.resolve.Resolve.gen_clauses
TOTAL    14.192         2 conda.resolve.Resolve.get_reduced_index
TOTAL   270.933         1 conda.resolve.Resolve.solve
```
Here, the most important parts are
```
TOTAL    62.973        62 conda.common.logic.CryptoMiniSatSolver.setup
TOTAL    24.026        62 conda.common.logic.PycoSatSolver.setup
```
and
```
TOTAL   119.002        62 conda.common.logic.CryptoMiniSatSolver.setup
TOTAL    39.833        62 conda.common.logic.PycoSatSolver.setup
```
. They show the total time taken for `62` invocations of `pycryptosat.Solver.add_clauses` with an `array.array("i")` or a `list` of `tuple`s of `int` as the input type for the set of clauses.
We see that
1. In this case the extra time taken to process the tuple list amounts to about half of the overall `add_clauses` runtime, and
2. The time taken by `pycosat`/PicoSAT is considerably slower.

----
*After* merging the latest improvements up to commit 10517a110f900ecea721b845fb76245be0b9fe6f, we get the following:
```
$ # after merging master (10517a1) + python-array + ClauseArray
$ conda create -nx --dry-run -v anaconda
Solving environment: ...working... === time_recorder total time and calls ===
TOTAL   112.682       853 conda.common.logic.Clauses.LinearBound_
TOTAL    99.651        62 conda.common.logic.Clauses._run_sat
TOTAL   208.951        12 conda.common.logic.Clauses.minimize
TOTAL    99.655        62 conda.common.logic.Clauses.sat
TOTAL     7.293        62 conda.common.logic.CryptoMiniSatSolver.invoke
TOTAL     1.587        62 conda.common.logic.CryptoMiniSatSolver.process_solution
TOTAL    58.094        62 conda.common.logic.CryptoMiniSatSolver.setup
TOTAL    23.862        62 conda.common.logic.PycoSatSolver.setup
TOTAL    70.517        62 conda.common.logic.SatSolver.run
TOTAL    20.656         2 conda.resolve.Resolve.gen_clauses
TOTAL    14.222         2 conda.resolve.Resolve.get_reduced_index
TOTAL   223.477         1 conda.resolve.Resolve.solve

$ # after merging master (10517a1) + python-array + ClauseList
$ conda create -nx --dry-run -v anaconda
Solving environment: ...working... === time_recorder total time and calls ===
TOTAL    81.224       853 conda.common.logic.Clauses.LinearBound_
TOTAL   150.507        62 conda.common.logic.Clauses._run_sat
TOTAL   229.594        12 conda.common.logic.Clauses.minimize
TOTAL   150.511        62 conda.common.logic.Clauses.sat
TOTAL     7.391        62 conda.common.logic.CryptoMiniSatSolver.invoke
TOTAL     1.659        62 conda.common.logic.CryptoMiniSatSolver.process_solution
TOTAL    92.382        62 conda.common.logic.CryptoMiniSatSolver.setup
TOTAL    40.174        62 conda.common.logic.PycoSatSolver.setup
TOTAL   104.823        62 conda.common.logic.SatSolver.run
TOTAL    19.895         2 conda.resolve.Resolve.gen_clauses
TOTAL    13.768         2 conda.resolve.Resolve.get_reduced_index
TOTAL   245.838         1 conda.resolve.Resolve.solve
```
, i.e.,
```
TOTAL    58.094        62 conda.common.logic.CryptoMiniSatSolver.setup
TOTAL    23.862        62 conda.common.logic.PycoSatSolver.setup
```
and
```
TOTAL    92.382        62 conda.common.logic.CryptoMiniSatSolver.setup
TOTAL    40.174        62 conda.common.logic.PycoSatSolver.setup
```
which shows a nearly 10 % improvement in the `array.array("i")` case (i.e., improvements to the core lib) and nearly 25 % improvements for the `List[Tuple[Int]]` case, i.e., also including improvements to the Python interface.

----
If we take the `pycosat` numbers as a benchmark, there is still some room for improvement -- but getting a 10 % time reduction already is nice!